refactor: Move DefaultRedirectionSection in separate folder
