+++ src/components/ProfileView.jsx
@@ -15,7 +15,7 @@ import Page from 'components/Page'
 import DeleteAccount from 'components/DeleteAccount'
 import Input from 'components/Input'
 import LanguageSection from 'components/LanguageSection'
-import DefaultRedirectionSection from 'components/DefaultRedirectionSection'
+import DefaultRedirectionSection from 'components/DefaultRedirectionSection/DefaultRedirectionSection'
 import TwoFA from 'components/2FA'
 import Import from 'components/Import'
 import ExportSection from 'components/export/ExportSection'
